[IIIF-170] Split Kakadu layers
Currently, we put all the Kakadu build stuff at /opt/kakadu in one layer.

Putting Kakadu libs in /opt/lib and bins in /opt/bin will make it easier on
Lambda functions that want to use the layer. Doing this, using the tool we
are though, means creating two different Lambda layers from our two Docker
layers. It's a little sad to create two Lambda layers (one for libs and one
for bins), but in the end is an okay trade-off because of the simplicity of
the tool that we're using.